Remove remaining file in core/operators/test
